# backend/docs/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "A personal blog system API built with Go and Gin",
    "title": "Blog API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "email": "support@blog.example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "paths": {
    "/articles": {
      "get": {
        "description": "Get all articles with optional filter by publication status",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "articles"
        ],
        "summary": "Get article list",
        "parameters": [
          {
            "type": "boolean",
            "description": "Filter by published status",
            "name": "published",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/model.ArticleSummaryResponse"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/articles/slug/{slug}": {
      "get": {
        "description": "Get a single article by its URL-friendly slug",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "articles"
        ],
        "summary": "Get article by slug",
        "parameters": [
          {
            "type": "string",
            "description": "Article slug",
            "name": "slug",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.ArticleResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/articles/{id}": {
      "get": {
        "description": "Get a single article by its numeric ID",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "articles"
        ],
        "summary": "Get article by ID",
        "parameters": [
          {
            "type": "integer",
            "description": "Article ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.ArticleResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "model.ArticleResponse": {
      "type": "object",
      "properties": {
        "author": {
          "$ref": "#/definitions/model.Author"
        },
        "category": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "coverImage": {
          "type": "string"
        },
        "excerpt": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "publishedAt": {
          "type": "string"
        },
        "readingTime": {
          "type": "integer"
        },
        "slug": {
          "type": "string"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "title": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
    "model.ArticleSummaryResponse": {
      "type": "object",
      "properties": {
        "category": {
          "type": "string"
        },
        "coverImage": {
          "type": "string"
        },
        "excerpt": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "publishedAt": {
          "type": "string"
        },
        "readingTime": {
          "type": "integer"
        },
        "slug": {
          "type": "string"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "title": {
          "type": "string"
        }
      }
    },
    "model.Author": {
      "type": "object",
      "properties": {
        "avatar": {
          "type": "string"
        },
        "bio": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    }
  }
}